requirements.txt
streamlit
openai 
pydantic
requests
python-dotenv
python-docx 
pypdf